Restructure project: organize files into folders and update references
- Organized files into a structured folder hierarchy for better clarity and maintainability.
- Updated file imports and references to ensure functionality after moving files.
- Verified compatibility across the project after the restructuring.
