+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__/
 *.pyc
 vscode/
 *.vscode
+*.pytest_cache
