+++ main.py
@@ -1,11 +1,11 @@
-from simulation import Simulation
-from renderer import Renderer
+from game.simulation import Simulation
+from game.renderer import Renderer
 
 if __name__ == "__main__":
     graphics = True
     simulation = Simulation()
     if graphics:
-        renderer = Renderer(simulation) 
+        renderer = Renderer(simulation)
         renderer.run()
     else:
         simulation.run()
