+++ map/obstacle.py
